+++ lib/chatterbox.dart
@@ -1,9 +1,10 @@
 library;
 
 export 'src/chatterbox.dart';
-export 'src/storage/dialog_store.dart';
 export 'src/dialog/flow.dart';
 export 'src/dialog/flow_impl.dart';
-export 'src/model/message_context.dart';
 export 'src/dialog/reaction.dart';
 export 'src/model/inline_button.dart';
+export 'src/model/message_context.dart';
+export 'src/storage/dialog_store.dart';
+export 'src/utils/hash_utils.dart';