+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-## 1.0.1
+## 1.0.2
+- Introduce toStepUri() function, that is used to reference between steps
+- Process callback for InlineButtons
 
+## 1.0.1
 - Add verification for InlineButtons
 - Process callback for InlineButtons
 
 ## 1.0.0
-
 - Initial version.